@@ -37,7 +37,10 @@ void main() {
 
         // Act
         final cameraId = await camera.createCamera(
-          CameraDescription(name: 'Test'),
+          CameraDescription(
+              name: 'Test',
+              lensDirection: CameraLensDirection.back,
+              sensorOrientation: 0),
           ResolutionPreset.high,
         );
 
@@ -48,7 +51,7 @@ void main() {
             arguments: {
               'cameraName': 'Test',
               'resolutionPreset': 'high',
-              'enableAudio': null
+              'enableAudio': true
             },
           ),
         ]);
@@ -70,7 +73,11 @@ void main() {
         // Act
         expect(
           () => camera.createCamera(
-            CameraDescription(name: 'Test'),
+            CameraDescription(
+              name: 'Test',
+              lensDirection: CameraLensDirection.back,
+              sensorOrientation: 0,
+            ),
             ResolutionPreset.high,
           ),
           throwsA(
@@ -97,7 +104,11 @@ void main() {
         // Act
         expect(
           () => camera.createCamera(
-            CameraDescription(name: 'Test'),
+            CameraDescription(
+              name: 'Test',
+              lensDirection: CameraLensDirection.back,
+              sensorOrientation: 0,
+            ),
             ResolutionPreset.high,
           ),
           throwsA(
@@ -122,7 +133,11 @@ void main() {
             });
         final camera = MethodChannelCamera();
         final cameraId = await camera.createCamera(
-          CameraDescription(name: 'Test'),
+          CameraDescription(
+            name: 'Test',
+            lensDirection: CameraLensDirection.back,
+            sensorOrientation: 0,
+          ),
           ResolutionPreset.high,
         );
 
@@ -165,7 +180,11 @@ void main() {
 
         final camera = MethodChannelCamera();
         final cameraId = await camera.createCamera(
-          CameraDescription(name: 'Test'),
+          CameraDescription(
+            name: 'Test',
+            lensDirection: CameraLensDirection.back,
+            sensorOrientation: 0,
+          ),
           ResolutionPreset.high,
         );
         Future<void> initializeFuture = camera.initializeCamera(cameraId);
@@ -197,8 +216,8 @@ void main() {
     });
 
     group('Event Tests', () {
-      MethodChannelCamera camera;
-      int cameraId;
+      late MethodChannelCamera camera;
+      late int cameraId;
       setUp(() async {
         MethodChannelMock(
           channelName: 'plugins.flutter.io/camera',
@@ -209,7 +228,11 @@ void main() {
         );
         camera = MethodChannelCamera();
         cameraId = await camera.createCamera(
-          CameraDescription(name: 'Test'),
+          CameraDescription(
+            name: 'Test',
+            lensDirection: CameraLensDirection.back,
+            sensorOrientation: 0,
+          ),
           ResolutionPreset.high,
         );
         Future<void> initializeFuture = camera.initializeCamera(cameraId);
@@ -352,8 +375,8 @@ void main() {
     });
 
     group('Function Tests', () {
-      MethodChannelCamera camera;
-      int cameraId;
+      late MethodChannelCamera camera;
+      int? cameraId;
       setUp(() async {
         MethodChannelMock(
           channelName: 'plugins.flutter.io/camera',
@@ -364,13 +387,17 @@ void main() {
         );
         camera = MethodChannelCamera();
         cameraId = await camera.createCamera(
-          CameraDescription(name: 'Test'),
+          CameraDescription(
+            name: 'Test',
+            lensDirection: CameraLensDirection.back,
+            sensorOrientation: 0,
+          ),
           ResolutionPreset.high,
         );
-        Future<void> initializeFuture = camera.initializeCamera(cameraId);
+        Future<void> initializeFuture = camera.initializeCamera(cameraId!);
         camera.cameraEventStreamController.add(
           CameraInitializedEvent(
-            cameraId,
+            cameraId!,
             1920,
             1080,
             ExposureMode.auto,
@@ -443,7 +470,7 @@ void main() {
             methods: {'takePicture': '/test/path.jpg'});
 
         // Act
-        XFile file = await camera.takePicture(cameraId);
+        XFile file = await camera.takePicture(cameraId!);
 
         // Assert
         expect(channel.log, <Matcher>[
@@ -478,7 +505,7 @@ void main() {
         );
 
         // Act
-        await camera.startVideoRecording(cameraId);
+        await camera.startVideoRecording(cameraId!);
 
         // Assert
         expect(channel.log, <Matcher>[
@@ -499,7 +526,7 @@ void main() {
 
         // Act
         await camera.startVideoRecording(
-          cameraId,
+          cameraId!,
           maxVideoDuration: Duration(seconds: 10),
         );
 
@@ -518,7 +545,7 @@ void main() {
         );
 
         // Act
-        XFile file = await camera.stopVideoRecording(cameraId);
+        XFile file = await camera.stopVideoRecording(cameraId!);
 
         // Assert
         expect(channel.log, <Matcher>[
@@ -537,7 +564,7 @@ void main() {
         );
 
         // Act
-        await camera.pauseVideoRecording(cameraId);
+        await camera.pauseVideoRecording(cameraId!);
 
         // Assert
         expect(channel.log, <Matcher>[
@@ -555,7 +582,7 @@ void main() {
         );
 
         // Act
-        await camera.resumeVideoRecording(cameraId);
+        await camera.resumeVideoRecording(cameraId!);
 
         // Assert
         expect(channel.log, <Matcher>[
@@ -573,10 +600,10 @@ void main() {
         );
 
         // Act
-        await camera.setFlashMode(cameraId, FlashMode.torch);
-        await camera.setFlashMode(cameraId, FlashMode.always);
-        await camera.setFlashMode(cameraId, FlashMode.auto);
-        await camera.setFlashMode(cameraId, FlashMode.off);
+        await camera.setFlashMode(cameraId!, FlashMode.torch);
+        await camera.setFlashMode(cameraId!, FlashMode.always);
+        await camera.setFlashMode(cameraId!, FlashMode.auto);
+        await camera.setFlashMode(cameraId!, FlashMode.off);
 
         // Assert
         expect(channel.log, <Matcher>[
@@ -599,8 +626,8 @@ void main() {
         );
 
         // Act
-        await camera.setExposureMode(cameraId, ExposureMode.auto);
-        await camera.setExposureMode(cameraId, ExposureMode.locked);
+        await camera.setExposureMode(cameraId!, ExposureMode.auto);
+        await camera.setExposureMode(cameraId!, ExposureMode.locked);
 
         // Assert
         expect(channel.log, <Matcher>[
@@ -619,8 +646,8 @@ void main() {
         );
 
         // Act
-        await camera.setExposurePoint(cameraId, Point<double>(0.5, 0.5));
-        await camera.setExposurePoint(cameraId, null);
+        await camera.setExposurePoint(cameraId!, Point<double>(0.5, 0.5));
+        await camera.setExposurePoint(cameraId!, null);
 
         // Assert
         expect(channel.log, <Matcher>[
@@ -647,7 +674,7 @@ void main() {
         );
 
         // Act
-        final minExposureOffset = await camera.getMinExposureOffset(cameraId);
+        final minExposureOffset = await camera.getMinExposureOffset(cameraId!);
 
         // Assert
         expect(minExposureOffset, 2.0);
@@ -666,7 +693,7 @@ void main() {
         );
 
         // Act
-        final maxExposureOffset = await camera.getMaxExposureOffset(cameraId);
+        final maxExposureOffset = await camera.getMaxExposureOffset(cameraId!);
 
         // Assert
         expect(maxExposureOffset, 2.0);
@@ -685,7 +712,7 @@ void main() {
         );
 
         // Act
-        final stepSize = await camera.getExposureOffsetStepSize(cameraId);
+        final stepSize = await camera.getExposureOffsetStepSize(cameraId!);
 
         // Assert
         expect(stepSize, 0.25);
@@ -704,7 +731,7 @@ void main() {
         );
 
         // Act
-        final actualOffset = await camera.setExposureOffset(cameraId, 0.5);
+        final actualOffset = await camera.setExposureOffset(cameraId!, 0.5);
 
         // Assert
         expect(actualOffset, 0.6);
@@ -724,8 +751,8 @@ void main() {
         );
 
         // Act
-        await camera.setFocusMode(cameraId, FocusMode.auto);
-        await camera.setFocusMode(cameraId, FocusMode.locked);
+        await camera.setFocusMode(cameraId!, FocusMode.auto);
+        await camera.setFocusMode(cameraId!, FocusMode.locked);
 
         // Assert
         expect(channel.log, <Matcher>[
@@ -744,8 +771,8 @@ void main() {
         );
 
         // Act
-        await camera.setFocusPoint(cameraId, Point<double>(0.5, 0.5));
-        await camera.setFocusPoint(cameraId, null);
+        await camera.setFocusPoint(cameraId!, Point<double>(0.5, 0.5));
+        await camera.setFocusPoint(cameraId!, null);
 
         // Assert
         expect(channel.log, <Matcher>[
@@ -766,7 +793,7 @@ void main() {
 
       test('Should build a texture widget as preview widget', () async {
         // Act
-        Widget widget = camera.buildPreview(cameraId);
+        Widget widget = camera.buildPreview(cameraId!);
 
         // Act
         expect(widget is Texture, isTrue);
@@ -791,7 +818,7 @@ void main() {
         );
 
         // Act
-        final maxZoomLevel = await camera.getMaxZoomLevel(cameraId);
+        final maxZoomLevel = await camera.getMaxZoomLevel(cameraId!);
 
         // Assert
         expect(maxZoomLevel, 10.0);
@@ -810,7 +837,7 @@ void main() {
         );
 
         // Act
-        final maxZoomLevel = await camera.getMinZoomLevel(cameraId);
+        final maxZoomLevel = await camera.getMinZoomLevel(cameraId!);
 
         // Assert
         expect(maxZoomLevel, 1.0);
@@ -829,7 +856,7 @@ void main() {
         );
 
         // Act
-        await camera.setZoomLevel(cameraId, 2.0);
+        await camera.setZoomLevel(cameraId!, 2.0);
 
         // Assert
         expect(channel.log, <Matcher>[
@@ -854,7 +881,7 @@ void main() {
 
         // Act & assert
         expect(
-            () => camera.setZoomLevel(cameraId, -1.0),
+            () => camera.setZoomLevel(cameraId!, -1.0),
             throwsA(isA<CameraException>()
                 .having((e) => e.code, 'code', 'ZOOM_ERROR')
                 .having((e) => e.description, 'description',
@@ -870,7 +897,7 @@ void main() {
 
         // Act
         await camera.lockCaptureOrientation(
-            cameraId, DeviceOrientation.portraitUp);
+            cameraId!, DeviceOrientation.portraitUp);
 
         // Assert
         expect(channel.log, <Matcher>[
@@ -887,7 +914,7 @@ void main() {
         );
 
         // Act
-        await camera.unlockCaptureOrientation(cameraId);
+        await camera.unlockCaptureOrientation(cameraId!);
 
         // Assert
         expect(channel.log, <Matcher>[
