@@ -17,7 +17,11 @@ enum CameraLensDirection {
 /// Properties of a camera device.
 class CameraDescription {
   /// Creates a new camera description with the given properties.
-  CameraDescription({this.name, this.lensDirection, this.sensorOrientation});
+  CameraDescription({
+    required this.name,
+    required this.lensDirection,
+    required this.sensorOrientation,
+  });
 
   /// The name of the camera device.
   final String name;
