@@ -4,6 +4,7 @@
 
 import 'package:camera_platform_interface/camera_platform_interface.dart';
 import 'package:camera_platform_interface/src/method_channel/method_channel_camera.dart';
+import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/mockito.dart';
 import 'package:plugin_platform_interface/plugin_platform_interface.dart';
@@ -117,7 +118,8 @@ void main() {
 
       // Act & Assert
       expect(
-        () => cameraPlatform.lockCaptureOrientation(1, null),
+        () => cameraPlatform.lockCaptureOrientation(
+            1, DeviceOrientation.portraitUp),
         throwsUnimplementedError,
       );
     });
@@ -155,7 +157,14 @@ void main() {
 
       // Act & Assert
       expect(
-        () => cameraPlatform.createCamera(null, null),
+        () => cameraPlatform.createCamera(
+          CameraDescription(
+            name: 'back',
+            lensDirection: CameraLensDirection.back,
+            sensorOrientation: 0,
+          ),
+          ResolutionPreset.high,
+        ),
         throwsUnimplementedError,
       );
     });
@@ -168,7 +177,7 @@ void main() {
 
       // Act & Assert
       expect(
-        () => cameraPlatform.initializeCamera(null),
+        () => cameraPlatform.initializeCamera(1),
         throwsUnimplementedError,
       );
     });
@@ -220,7 +229,7 @@ void main() {
 
       // Act & Assert
       expect(
-        () => cameraPlatform.setFlashMode(1, null),
+        () => cameraPlatform.setFlashMode(1, FlashMode.auto),
         throwsUnimplementedError,
       );
     });
@@ -233,7 +242,7 @@ void main() {
 
       // Act & Assert
       expect(
-        () => cameraPlatform.setExposureMode(1, null),
+        () => cameraPlatform.setExposureMode(1, ExposureMode.auto),
         throwsUnimplementedError,
       );
     });
@@ -298,7 +307,7 @@ void main() {
 
       // Act & Assert
       expect(
-        () => cameraPlatform.setExposureOffset(1, null),
+        () => cameraPlatform.setExposureOffset(1, 2.0),
         throwsUnimplementedError,
       );
     });
@@ -311,7 +320,7 @@ void main() {
 
       // Act & Assert
       expect(
-        () => cameraPlatform.setFocusMode(1, null),
+        () => cameraPlatform.setFocusMode(1, FocusMode.auto),
         throwsUnimplementedError,
       );
     });
